Move LED BLE coordinator to separate module
